+++ tolk/pipe-discover-parse-sources.cpp
@@ -1,0 +1,62 @@
+/*
+    This file is part of TON Blockchain source code.
+
+    TON Blockchain is free software; you can redistribute it and/or
+    modify it under the terms of the GNU General Public License
+    as published by the Free Software Foundation; either version 2
+    of the License, or (at your option) any later version.
+
+    TON Blockchain is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with TON Blockchain.  If not, see <http://www.gnu.org/licenses/>.
+
+    In addition, as a special exception, the copyright holders give permission
+    to link the code of portions of this program with the OpenSSL library.
+    You must obey the GNU General Public License in all respects for all
+    of the code used other than OpenSSL. If you modify file(s) with this
+    exception, you may extend this exception to your version of the file(s),
+    but you are not obligated to do so. If you do not wish to do so, delete this
+    exception statement from your version. If you delete this exception statement
+    from all source files in the program, then also delete it here.
+*/
+#include "tolk.h"
+#include "ast.h"
+#include "ast-from-tokens.h"
+#include "compiler-state.h"
+
+namespace tolk {
+
+AllSrcFiles pipeline_discover_and_parse_sources(const std::string& stdlib_filename, const std::string& entrypoint_filename) {
+  G.all_src_files.locate_and_register_source_file(stdlib_filename, {});
+  G.all_src_files.locate_and_register_source_file(entrypoint_filename, {});
+
+  while (SrcFile* file = G.all_src_files.get_next_unparsed_file()) {
+    tolk_assert(!file->ast);
+
+    file->ast = parse_src_file_to_ast(file);
+
+    for (AnyV v_toplevel : file->ast->as<ast_tolk_file>()->get_toplevel_declarations()) {
+      if (auto v_include = v_toplevel->try_as<ast_include_statement>()) {
+        size_t pos = file->rel_filename.rfind('/');
+        std::string rel_filename = pos == std::string::npos
+          ? v_include->get_file_name()
+          : file->rel_filename.substr(0, pos + 1) + v_include->get_file_name();
+
+        SrcFile* imported = G.all_src_files.locate_and_register_source_file(rel_filename, v_include->loc);
+        file->imports.push_back(SrcFile::ImportStatement{imported});
+        v_include->mutate_set_src_file(imported);
+      }
+    }
+  }
+
+  // todo #ifdef TOLK_PROFILING
+  // lexer_measure_performance(G.all_src_files.get_all_files());
+
+  return G.all_src_files.get_all_files();
+}
+
+} // namespace tolk
